requirements.txt
mysql-connector==2.2.9
tornado==6.0.4
